moved logic to service folder
