+++ src/controllers/user.js
@@ -1,7 +1,4 @@
-/* eslint-disable camelcase */
-const bcrypt = require('bcrypt');
-const registerValidation = require('../validation/user');
-const User = require('../services/user');
+const registerService = require('../services/user');
 const { successResponse, handleError, errorResponse } = require('../utils/response');
 
 /**
@@ -17,19 +14,7 @@ module.exports = class UserController {
    */
   static async createUser(req, res) {
     try {
-      const { error } = registerValidation(req.body);
-      if (error) {
-        return res.status(400).json({ status: 400, error: error.message });
-      }
-      const { email, first_name, last_name, password, phone_number } = req.body;
-      const Email = email.toLowerCase();
-      const emailExist = await User.emailExist(Email);
-      if (emailExist) return errorResponse(res, 409, 'Email already used by another user.');
-      const phoneExist = await User.phoneExist(phone_number);
-      if (phoneExist) return errorResponse(res, 409, 'Phone already used by another user.');
-      const hashedPassword = await bcrypt.hash(password, 10);
-      const newUser = { email: Email, first_name, last_name, password: hashedPassword, phone_number };
-      const createdUser = await User.createUser(newUser);
+      const createdUser = await registerService(req, res);
       return successResponse(res, 200, 'User Created successfully!', createdUser);
     } catch (error) {
       handleError(error, req);
